feat: duplicated MainController removed
